Make it as a link wrapper instead
