+++ apps/web/components/index.ts
@@ -12,6 +12,7 @@ export * from './SideMenu/SideMenu'
 export * from './EmailSignup/EmailSignup'
 export * from './StoryList/StoryList'
 export * from './Content/Content'
+export * from './ContentLink/ContentLink'
 export * from './LatestNewsSection'
 export * from './NewsCard'
 export * from './LogoList/LogoList'
